finish off permissions

+++ server/api/event/views.py
@@ -1,21 +1,21 @@
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 
-from django.http import HttpResponse
-
 from .models import Event
 from .serializers import EventSerialiser
 
 
 @api_view(['POST'])
 def create_event(request):
-    serialiser = EventSerialiser(data=request.data)
-    if serialiser.is_valid():
-        serialiser.save()
-        return Response(serialiser.data, status=200)
+    if request.user.is_authenticated is False:
+        return Response("User not authenticated", status=401)
     else:
-        # print(serialiser)
-        return Response(serialiser.errors, status=400)
+        serialiser = EventSerialiser(data=request.data)
+        if serialiser.is_valid():
+            serialiser.save()
+            return Response(serialiser.data, status=200)
+        else:
+            return Response(serialiser.errors, status=400)
 
 
 @api_view(['GET'])
@@ -28,7 +28,7 @@ def get_event(request, event_id):
 
 @api_view(['GET'])
 def get_events(request):
-    if request.user.is_authenticated:
+    if request.user.is_authenticated is True:
         if (request.user.team_id is not None):
             events = Event.objects.filter(team_id=request.user.team_id)
             team_serializer = EventSerialiser(events, many=True)
@@ -36,7 +36,7 @@ def get_events(request):
         else:
             events = Event.objects.all()
             serializer = EventSerialiser(events, many=True)
-            return HttpResponse(serializer.data)
+            return Response(serializer.data)
     else:
         events = Event.objects.filter(is_public=True)
         limited_serializer = EventSerialiser(events, many=True)
@@ -45,22 +45,39 @@ def get_events(request):
 
 @api_view(['PUT'])
 def update_event(request, event_id):
-    event = Event.objects.get(event_id=event_id)
-    print(event)
-    if event.is_public is False:
-        serializer = EventSerialiser(instance=event, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-        return Response(serializer.errors)
+    if request.user.is_authenticated is False:
+        return Response("User not authenticated", status=401)
     else:
-        return Response("Event is not private")
+        if request.user.team_admin is False:
+            return Response("User is not authorised to update an event", status=403)
+        else:
+            event = Event.objects.get(event_id=event_id)
+            if request.user.team_id == event.team_id:
+                if event.is_public is False:
+                    serializer = EventSerialiser(instance=event, data=request.data)
+                    if serializer.is_valid():
+                        serializer.save()
+                    return Response(serializer.errors, status=400)
+                else:
+                    return Response("Event is not private", status=401)
+            else:
+                return Response("User is not authorised to update this event", status=401)
 
 
 @api_view(['DELETE'])
 def delete_event(request, event_id):
-    event = Event.objects.get(event_id=event_id)
-    if event.is_public is False:
-        event.delete()
-        return Response("Event successfully deleted")
+    if request.user.is_authenticated is False:
+        return Response("User not authenticated", status=401)
     else:
-        return Response("Event is not private")
+        if request.user.team_admin is False:
+            return Response("User is not authorised to update an event", status=403)
+        else:
+            event = Event.objects.get(event_id=event_id)
+            if request.user.team_id == event.team_id:
+                if event.is_public is False:
+                    event.delete()
+                    return Response("Event successfully deleted")
+                else:
+                    return Response("Event is not private")
+            else:
+                return Response("User is not authorised to update this event", status=401)
